requirements.txt
numpy~=1.26.4
pandas~=2.2.1
pillow~=10.2.0
matplotlib~=3.8.3